# ch02/mux-router/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Gorilla Mux Article API",
    "description": "A simple RESTful API demonstrating path-based and query-based matching using Gorilla Mux.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8000",
      "description": "Local development server"
    }
  ],
  "paths": {
...
    "/query-based/articles": {
      "get": {
        "summary": "Query-based article retrieval",
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string",
                  "example": "Got parameter id: 123!\nGot parameter category: books!"
                }
              }
            }
          }
        }
      }
    },
...
  }
}